+++ tests/test_split.py
@@ -1,0 +1,69 @@
+import unittest
+
+from .helper import equals
+
+import gpxpy.parser as mod_parser
+
+
+class TestSplit(unittest.TestCase):
+    def test_split_on_impossible_index(self):
+        f = open('test_files/cerknicko-jezero.gpx')
+        parser = mod_parser.GPXParser(f)
+        gpx = parser.parse()
+        f.close()
+
+        track = gpx.tracks[0]
+
+        before = len(track.segments)
+        track.split(1000, 10)
+        after = len(track.segments)
+
+        self.assertTrue(before == after)
+
+    def test_split(self):
+        f = open('test_files/cerknicko-jezero.gpx')
+        parser = mod_parser.GPXParser(f)
+        gpx = parser.parse()
+        f.close()
+
+        track = gpx.tracks[1]
+
+        track_points_no = track.get_points_no()
+
+        before = len(track.segments)
+        track.split(0, 10)
+        after = len(track.segments)
+
+        self.assertTrue(before + 1 == after)
+        print('Points in first (split) part:', len(track.segments[0].points))
+
+        # From 0 to 10th point == 11 points:
+        self.assertTrue(len(track.segments[0].points) == 11)
+        self.assertTrue(len(track.segments[0].points) + len(track.segments[1].points) == track_points_no)
+
+        # Now split the second track
+        track.split(1, 20)
+        self.assertTrue(len(track.segments[1].points) == 21)
+        self.assertTrue(len(track.segments[0].points) + len(track.segments[1].points) + len(track.segments[2].points) == track_points_no)
+
+    def test_split_and_join(self):
+        f = open('test_files/cerknicko-jezero.gpx')
+        parser = mod_parser.GPXParser(f)
+        gpx = parser.parse()
+        f.close()
+
+        track = gpx.tracks[1]
+
+        original_track = track.clone()
+
+        track.split(0, 10)
+        track.split(1, 20)
+
+        self.assertTrue(len(track.segments) == 3)
+        track.join(1)
+        self.assertTrue(len(track.segments) == 2)
+        track.join(0)
+        self.assertTrue(len(track.segments) == 1)
+
+        # Check that this split and joined track is the same as the original one:
+        self.assertTrue(equals(track, original_track))
